+++ ChessUI/Time Control/TimeControlOptions.cs
@@ -2,9 +2,9 @@
 {
     public class TimeControlOptions
     {
-        public int WhiteInitialTimeMs { get; set; } = 300;
-        public int BlackInitialTimeMs { get; set; } = 300;
-        public int WhiteIncrementMs { get; set; } = 0;
-        public int BlackIncrementMs { get; set; } = 0;
+        public int WhiteInitialTimeMs { get; set; } = 300_000;
+        public int BlackInitialTimeMs { get; set; } = 300_000;
+        public int WhiteIncrementMs { get; set; } = 5_000;
+        public int BlackIncrementMs { get; set; } = 5_000;
     }
 }
