+++ ChessUI/Views/GameSettings.xaml.cs
@@ -34,19 +34,19 @@ namespace ChessUI.Views
         }
         private void WhiteTimeChanged(object sender, TextChangedEventArgs e)
         {
-            GameSetup.TimeControl.WhiteInitialTimeMs = Int32.Parse(WhiteTime.Text);
+            GameSetup.TimeControl.WhiteInitialTimeMs = Int32.Parse(WhiteTime.Text) * 1000;
         }
         private void BlackTimeChanged(object sender, TextChangedEventArgs e)
         {
-            GameSetup.TimeControl.BlackInitialTimeMs = Int32.Parse(BlackTime.Text);
+            GameSetup.TimeControl.BlackInitialTimeMs = Int32.Parse(BlackTime.Text) * 1000;
         }
         private void WhiteIncrementChanged(object sender, TextChangedEventArgs e)
         {
-            GameSetup.TimeControl.WhiteIncrementMs = Int32.Parse(WhiteIncrement.Text);
+            GameSetup.TimeControl.WhiteIncrementMs = Int32.Parse(WhiteIncrement.Text) * 1000;
         }
         private void BlackIncrementChanged(object sender, TextChangedEventArgs e)
         {
-            GameSetup.TimeControl.BlackIncrementMs = Int32.Parse(BlackIncrement.Text);
+            GameSetup.TimeControl.BlackIncrementMs = Int32.Parse(BlackIncrement.Text) * 1000;
         }
 
         private void ConfirmBtn_Click(object sender, RoutedEventArgs e)
