+++ ChessUI/Views/PromotionSelection.xaml
@@ -1,0 +1,38 @@
+﻿<Window x:Class="ChessUI.PromotionSelection"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:ChessUI"
+        mc:Ignorable="d"
+        Title="PromotionSelection" Height="430" Width="430" FontSize="30">
+    <Grid>
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="30" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="30" />
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="30" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="10" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="10" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="10" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="30" />
+        </Grid.RowDefinitions>
+
+
+        <Button Name="Queen" Grid.Row="1" Grid.Column="1" Click ="QueenClick" Background="DarkSeaGreen" Content="Queen" />
+
+        <Button Name="Rook" Grid.Row="3" Grid.Column="1" Click ="RookClick" Background="DarkSeaGreen" Content="Rook" />
+
+        <Button Name="Bishop" Grid.Row="5" Grid.Column="1" Click ="BishopClick" Background="DarkSeaGreen" Content="Bishop" />
+
+        <Button Name="Knight" Grid.Row="7" Grid.Column="1" Click ="KnightClick" Background="DarkSeaGreen" Content="Knight" />
+    </Grid>
+</Window>
