--- ChessUI/start.txt
@@ -1,1 +1,0 @@
-rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1
